Update date only if there actually is a change
Calling this.set('date', date) with a cloned object leads to infinite recursion (which is broken by the browser once the maximum stack size is reached). 

I did not completely follow the Polymer code, but it looks like Polymer keeps calling _dateChanged() with alternating oldValue and newValue. The problem is that it cannot determine equality between them itself as they are objects.