Fix the blank page displayed after the max bound when a max-date is defined
Before this commit, when a max-date was defined, you were able to swipe
to the next month even after the max bound, this caused an issue because
it displayed a blank page.
Now when you reached the max month you can’t swipe another time, you
stay on this month.

Fixes #114