Update deprecated @apply and var() CSS syntax
Polymer 2.0 is production ready, and we advise all projects to begin migration. To aid in that migration, we're writing and applying some automated fixes for the mechanical parts of the upgrade.

This change migrates uses of @apply to remove the deprecated parentheses syntax.

These changes are backwards-compatible with Polymer 1.0 applications.